+++ checkpoint/orbax/checkpoint/experimental/v1/_src/path/async_utils.py
@@ -14,86 +14,229 @@
 
 """Utilities for processing paths in asynchronous contexts."""
 
+from __future__ import annotations
+
 import asyncio
-from typing import Iterable
+import time
+from typing import Awaitable, Iterable, Sequence
+
+from absl import logging
+import jax
 from orbax.checkpoint._src.futures import future
 from orbax.checkpoint._src.futures import synchronization
-from orbax.checkpoint._src.path import atomicity
+from orbax.checkpoint._src.path import async_path
 from orbax.checkpoint._src.path import atomicity_types
 from orbax.checkpoint.experimental.v1._src.context import context as context_lib
 from orbax.checkpoint.experimental.v1._src.path import types
+from orbax.checkpoint.experimental.v1._src.synchronization import multihost
 
 Path = types.Path
 PathLike = types.PathLike
-PathAwaitingCreation = types.PathAwaitingCreation
+TemporaryPath = atomicity_types.TemporaryPath
 
 
-class _PathAwaitingCreation(PathAwaitingCreation):
-  """Implementation of :py:class:`.PathAwaitingCreation` that wraps an awaitable."""
+async def _create_paths(
+    tmp_path: TemporaryPath,
+    *,
+    subdirectories: Iterable[str],
+    context: context_lib.Context,
+    operation_id: str,
+    completion_signals: Sequence[synchronization.HandlerAwaitableSignal],
+):
+  """Creates :py:class:`.`TemporaryPath` and subdirectories."""
+  active_processes = context.multiprocessing_options.active_processes
+  primary_host = context.multiprocessing_options.primary_host
+  barrier_sync_key_prefix = (
+      context.multiprocessing_options.barrier_sync_key_prefix
+  )
+  if multihost.is_primary_host(primary_host):
+    start = time.time()
+    path = await tmp_path.create()
+    # subdirectory assumed to not have any nesting.
+    subdir_ops = [
+        async_path.mkdir(path / subdirectory, parents=False, exist_ok=False)
+        for subdirectory in subdirectories
+    ]
+    await asyncio.gather(*subdir_ops)
+    directory_creation_secs = time.time() - start
+    jax.monitoring.record_event_duration_secs(
+        '/jax/orbax/write/directory_creation_secs',
+        directory_creation_secs,
+    )
+    jax.monitoring.record_event_duration_secs(
+        '/jax/orbax/write/async_directory_creation_secs',
+        directory_creation_secs,
+    )
+    logging.vlog(
+        1,
+        'Asynchronous directory creation took %s seconds',
+        directory_creation_secs,
+    )
+    future.set_signals(completion_signals, operation_id=operation_id)
+  await multihost.sync_global_processes(
+      multihost.unique_barrier_key(
+          'create_directory:post',
+          prefix=barrier_sync_key_prefix,
+      ),
+      operation_id=operation_id,
+      timeout=multihost.coordination_timeout(),
+      processes=active_processes,
+  )
 
-  def __init__(
-      self, directory: Path, f: future.Future
-  ):
-    self._directory = directory
-    self._f = f
 
-  def __truediv__(
-      self, other: PathAwaitingCreation | PathLike
-  ) -> PathAwaitingCreation:
-    if isinstance(other, PathAwaitingCreation):
-      other = other.path
-    return _PathAwaitingCreation(self.path / other, self._f)
+class _SubdirectoryAwaitingCreation(types.PathAwaitingCreation):
+  """A :py:class:`.`PathAwaitingCreation` that is a subdirectory of a base path.
+
+  It expects to receive a base :py:class:`.PathAwaitingCreation` during
+  initialization, and
+  will wait for that path to be created before allowing access to the
+  subdirectory. Any further subdirectory appending will still reference the
+  same base path.
+  """
+
+  def __init__(self, path: Path, base_path: types.PathAwaitingCreation):
+    self._base_path = base_path
+    self._path = path
+
+  def __truediv__(self, other: PathLike) -> types.PathAwaitingCreation:
+    if not isinstance(other, PathLike):
+      raise TypeError(f'Expected PathLike, got {type(other)}.')
+    return _SubdirectoryAwaitingCreation(
+        self.path / other,
+        self._base_path,
+    )
 
   @property
   def path(self) -> Path:
-    return self._directory
+    return self._path
 
   async def await_creation(self) -> Path:
-    await asyncio.to_thread(self._f.result)
-    return self._directory
+    """Waits for the directory to be created."""
+    await self._base_path.await_creation()
+    return self._path
 
 
-def start_async_mkdir(
-    path: atomicity_types.TemporaryPath,
-    subdirectories: Iterable[str] = (),
-) -> PathAwaitingCreation:
-  """Starts async directory creation on a :py:class:`~.atomicity_types.TemporaryPath`.
+class PathAwaitingCreation(types.PathAwaitingCreation):
+  """Implementation of :py:class:`.PathAwaitingCreation` that creates paths asynchronously.
 
-  The `mkdir` operation starts immediately in a background thread on `path`.
-  Creation also starts for any provided subdirectories. A
-  :py:class:`.PathAwaitingCreation` object is returned.
+  This implementation also includes a `create` API that allows a caller to
+  create the paths immediately. The `await_creation` API may also trigger a
+  creation if it has not already started. This allows implementers of
+  custom handlers to trigger themselves without running into hangs.
 
-  Subsequent operations on the returned object will NOT create any additional
-  directories. For example, using::
+  Creation is carried out uniquely. Repeated calls to `create` will NOT create
+  any additional directories. For example, using::
 
-    p = start_async_mkdir(path, ['a', 'b'])
+    p = PathAwaitingCreation.build(path, ['a', 'b'])
     new_p = p / 'c'
 
   will not create a subdirectory named 'c'. Only directories `path` and `path/a`
   and `path/b` will be created.
 
-  Args:
-    path: The path to create. May be an instance of
-      :py:class:`~.atomicity_types.TemporaryPath`.
-    subdirectories: A sequence of subdirectories to create under `path`.
-
-  Returns:
-    A :py:class:`.PathAwaitingCreation` object.
+  For async directory creations, `create` should be called from a background
+  thread. We allow the background thread to initiate the mkdir operation
+  directly, instead of initiating in the main thread and carrying it over to the
+  background thread.
+  COST:
+   - Introduces a small amount of slowdown, but the overall impact should be
+     very marginal given that it would affect only the background thread.
+  BENEFIT:
+   - Simplifies the logic involved in allowing control of the async operations
+     to be transferred between threads.
+  We should not go all the way to this proposal's logical conclusion and just
+  require individual handlers to create their own directories. Remember that
+  directories must be created in a centralized place to avoid duplicate
+  requests, which creates additional QPS burden on the filesystem.
   """
-  context = context_lib.get_context()
 
-  # TODO(b/407609827): V0 TypeHandler implementations, which are still used on
-  # the saving path, do not have knowledge of the `PathAwaitingCreation`, and
-  # instead rely on signals. We will need to continue using signals for now,
-  # until `LeafHandler` implementations can be updated.
-  completion_signals = [
-      synchronization.HandlerAwaitableSignal.STEP_DIRECTORY_CREATION,
-      synchronization.HandlerAwaitableSignal.ITEM_DIRECTORY_CREATION,
-  ]
-  f = atomicity.create_all_async(
-      [path],
-      completion_signals=completion_signals,
-      multiprocessing_options=context.multiprocessing_options.v0(),
-      subdirectories=[name for name in subdirectories],
-  )
-  return _PathAwaitingCreation(path.get(), f)
+  def __init__(
+      self,
+      path: Path,
+      awaitable: Awaitable[None],
+  ):
+    self._path = path
+    self._awaitable = awaitable
+    self._creation_completed = False
+    self._lock = asyncio.Lock()
+
+  def __truediv__(self, other: PathLike) -> types.PathAwaitingCreation:
+    """Creates a new :py:class:`.PathAwaitingCreation` that appends to the current path.
+
+    The path returned will reference the original "base" object, and
+    `await_creation` will wait for the base path to be created. Subdirectories
+    of the base path appended in this way (unless they were already created by
+    the `create` operation) will not be automatically created, and the user will
+    need to create them after calling `await_creation`.
+
+    Args:
+      other: The subdirectory to create.
+
+    Returns:
+      A new :py:class:`.PathAwaitingCreation` that appends to the current path.
+    """
+    if not isinstance(other, PathLike):
+      raise TypeError(f'Expected PathLike, got {type(other)}.')
+    return _SubdirectoryAwaitingCreation(
+        self.path / other,
+        self,
+    )
+
+  @property
+  def path(self) -> Path:
+    return self._path
+
+  async def create(self) -> Path:
+    """Creates the directory if it has not already been created.
+
+    Use a lock to ensure that only one process starts the creation.
+    Any other processes will wait at the lock if a creation is already in
+    progress, then return immediately since `_creation_completed` will have
+    already been set.
+
+    Returns:
+      The path that was created.
+    """
+    async with self._lock:
+      if self._creation_completed:
+        return self._path
+      self._creation_completed = True
+      await self._awaitable
+      # Any operations after this might cause deadlock if self._awaitable raises
+      # an exception.
+    return self._path
+
+  async def await_creation(self) -> Path:
+    """Waits for the directory to be created.
+
+    Creation will be triggered if it has not already started.
+    Returns:
+      The path that was created.
+    """
+    await self.create()  # This is a no-op if already created.
+    return self._path
+
+  @classmethod
+  def build(
+      cls,
+      path: TemporaryPath,
+      subdirectories: Iterable[str],
+  ) -> PathAwaitingCreation:
+    # TODO(b/407609827): V0 TypeHandler implementations, which are still used on
+    # the saving path, do not have knowledge of the `PathAwaitingCreation`, and
+    # instead rely on signals. We will need to continue using signals for now,
+    # until `LeafHandler` implementations can be updated.
+    completion_signals = [
+        synchronization.HandlerAwaitableSignal.STEP_DIRECTORY_CREATION,
+        synchronization.HandlerAwaitableSignal.ITEM_DIRECTORY_CREATION,
+    ]
+    future.AwaitableSignalsContract.add_to_awaitable_signals_contract(
+        completion_signals
+    )
+    awaitable = _create_paths(
+        path,
+        subdirectories=subdirectories,
+        context=context_lib.get_context(),
+        operation_id=context_lib.get_context().operation_id(),
+        completion_signals=completion_signals,
+    )
+    return cls(path.get(), awaitable)