+++ checkpoint/orbax/checkpoint/experimental/v1/_src/saving/execution.py
@@ -14,6 +14,8 @@
 
 """Internal utilities for saving whole and partial checkpoints."""
 
+from __future__ import annotations
+
 import asyncio
 import hashlib
 import time
@@ -46,6 +48,16 @@ from orbax.checkpoint.experimental.v1._src.tree import types as tree_types
 InternalCheckpointMetadata = (
     step_metadata_serialization.InternalCheckpointMetadata
 )
+AsyncResponse = async_types.AsyncResponse
+
+
+def _should_create_directories_synchronously(
+    context: context_lib.Context, partial_save: bool
+):
+  return (
+      partial_save
+      or not context.async_options.create_directories_asynchronously
+  )
 
 
 def add_internal_checkpointables(
@@ -70,13 +82,13 @@ def add_internal_checkpointables(
   return checkpointables
 
 
-class SaveResponse(async_types.AsyncResponse[None]):
-  """An `AsyncResponse` representing the result of :py:func:`.save_pytree_async`."""
+class _SaveResponse(AsyncResponse[None]):
+  """An :py:class:`.AsyncResponse` representing the result of:py:func:`.save_pytree_async`."""
 
   def __init__(
       self,
       operation_id: str,
-      tmp_path: atomicity_types.TemporaryPath,
+      temporary_path: _TemporaryPathAwaitingCreation,
       handler_typestrs: dict[str, str],
       background_awaitable: Awaitable[None],
       *,
@@ -86,7 +98,7 @@ class SaveResponse(async_types.AsyncResponse[None]):
       async_origin: bool,
   ):
     self._operation_id = operation_id
-    self._tmp_path = tmp_path
+    self._temporary_path = temporary_path
     self._handler_typestrs = handler_typestrs
     self._background_awaitable = background_awaitable
     self._start_time = start_time
@@ -97,11 +109,66 @@ class SaveResponse(async_types.AsyncResponse[None]):
         self._finalize_save()
     )
 
+  @classmethod
+  def create(
+      cls,
+      background_awaitable: Awaitable[None],
+      checkpointables: dict[str, Any],
+      temporary_path: _TemporaryPathAwaitingCreation,
+      start_time: float,
+      *,
+      context: context_lib.Context,
+      custom_metadata: tree_types.JsonType | None,
+      async_origin: bool,
+  ) -> _SaveResponse:
+    """Creates and returns the final AsyncResponse for a save operation."""
+    blocking_duration_secs = time.time() - start_time
+    jax.monitoring.record_event_duration_secs(
+        '/jax/checkpoint/write/async/blocking_duration_secs',
+        blocking_duration_secs,
+    )
+    logging.info(
+        'Finished blocking save in %.2f seconds. Continuing to write to %s.',
+        blocking_duration_secs,
+        temporary_path.temporary_path.get_final(),
+    )
+
+    handler = composite_handler.CompositeHandler(
+        context.checkpointables_options.registry
+    )
+    handler_typestrs = {
+        name: handler_types.typestr(type(handler))
+        for name, handler in handler.get_handlers_for_save(
+            checkpointables
+        ).items()
+    }
+
+    return cls(
+        context.operation_id(),
+        temporary_path,
+        handler_typestrs,
+        background_awaitable,
+        start_time=start_time,
+        custom_metadata=custom_metadata,
+        context=context,
+        async_origin=async_origin,
+    )
+
   async def _finalize_save(self):
     logging.info(
         '[process=%s] Finalizing checkpoint on %s',
         multihost.process_index(),
-        self._tmp_path.get(),
+        self._temporary_path.temporary_path.get(),
+    )
+    logging.info(
+        '[process=%s] Creating directories on %s',
+        multihost.process_index(),
+        self._temporary_path.temporary_path.get(),
+    )
+    await self._temporary_path.path_awaiting_creation.create()
+    logging.info(
+        '[process=%s] Waiting for background save operations',
+        multihost.process_index(),
     )
     await self._background_awaitable
     logging.vlog(
@@ -126,7 +193,7 @@ class SaveResponse(async_types.AsyncResponse[None]):
       )
       await metadata_serialization.write(
           metadata_serialization.checkpoint_metadata_file_path(
-              self._tmp_path.get()
+              self._temporary_path.temporary_path.get()
           ),
           internal_metadata.serialize(),
       )
@@ -136,7 +203,7 @@ class SaveResponse(async_types.AsyncResponse[None]):
           multihost.process_index(),
       )
       await atomicity.on_commit_callback(
-          self._tmp_path,
+          self._temporary_path.temporary_path,
           checkpoint_start_time=self._start_time,
       )
 
@@ -157,7 +224,7 @@ class SaveResponse(async_types.AsyncResponse[None]):
     )
     total_duration_secs = time.time() - self._start_time
     event_tracking.record_save_completion(
-        self._tmp_path.get_final(),
+        self._temporary_path.temporary_path.get_final(),
         total_duration_secs=total_duration_secs,
         async_origin=self._async_origin,
     )
@@ -166,10 +233,9 @@ class SaveResponse(async_types.AsyncResponse[None]):
     return self._thread_runner.result(timeout=timeout)
 
 
-async def run_blocking_save(
-    tmp_path: atomicity_types.TemporaryPath,
+async def _run_blocking_save(
+    temporary_path: _TemporaryPathAwaitingCreation,
     checkpointables: dict[str, Any],
-    subdirectories: Iterable[str],
     *,
     overwrite: bool,
     context: context_lib.Context,
@@ -180,9 +246,8 @@ async def run_blocking_save(
   This includes directory creation and calling the handler's save method.
 
   Args:
-    tmp_path: The temporary path to save the checkpointables to.
+    temporary_path: The temporary path to save the checkpointables to.
     checkpointables: A mapping from checkpointable name to checkpointable.
-    subdirectories: A list of subdirectories to create under `tmp_path`.
     overwrite: Whether to overwrite an existing checkpoint in `tmp_path`.
     context: The context to use for the save operation.
     partial_save: Whether to save the checkpoint in partial mode.
@@ -191,26 +256,21 @@ async def run_blocking_save(
     An awaitable that will be completed when the synchronous portion of the save
     operation is complete.
   """
-  await context_lib.synchronize_next_operation_id()
-
   if not partial_save:
     await saving_path_utils.maybe_overwrite_existing(
-        tmp_path.get_final(), overwrite=overwrite, context=context
+        temporary_path.temporary_path.get_final(),
+        overwrite=overwrite,
+        context=context,
     )
 
   handler = composite_handler.CompositeHandler(
       context.checkpointables_options.registry
   )
-
-  # Directory creation is handled here.
-  tmp_path_awaiting_creation = path_async_utils.start_async_mkdir(
-      tmp_path, subdirectories
-  )
   if (
       partial_save
       or not context.async_options.create_directories_asynchronously
   ):
-    await tmp_path_awaiting_creation.await_creation()
+    await temporary_path.path_awaiting_creation.create()
 
   if partial_save:
     await multihost.sync_global_processes(
@@ -224,73 +284,16 @@ async def run_blocking_save(
 
   # Delegate to the handler to get the background awaitable.
   background_awaitable = await handler.save(
-      tmp_path_awaiting_creation, checkpointables
+      temporary_path.path_awaiting_creation, checkpointables
   )
   # Log write event for the final path.
-  event_tracking.record_write_event(tmp_path.get_final())
+  event_tracking.record_write_event(temporary_path.temporary_path.get_final())
 
   return background_awaitable
 
 
-def create_save_response(
-    background_awaitable: Awaitable[None],
-    checkpointables: dict[str, Any],
-    tmp_path: atomicity_types.TemporaryPath,
-    start_time: float,
-    *,
-    context: context_lib.Context,
-    custom_metadata: tree_types.JsonType | None,
-    async_origin: bool,
-) -> async_types.AsyncResponse[None]:
-  """Creates and returns the final AsyncResponse for a save operation.
-
-  Args:
-    background_awaitable: The awaitable for the background save operation.
-    checkpointables: A dictionary of checkpointables.
-    tmp_path: The temporary path used for saving.
-    start_time: The time when the save operation started.
-    context: The Orbax context.
-    custom_metadata: Optional custom metadata to include in the checkpoint.
-    async_origin: Whether the save operation originated from an async call.
-
-  Returns:
-    An AsyncResponse for the save operation.
-  """
-  blocking_duration_secs = time.time() - start_time
-  jax.monitoring.record_event_duration_secs(
-      '/jax/checkpoint/write/async/blocking_duration_secs',
-      blocking_duration_secs,
-  )
-  logging.info(
-      'Finished blocking save in %.2f seconds. Continuing to write to %s.',
-      blocking_duration_secs,
-      tmp_path.get_final(),
-  )
-
-  handler = composite_handler.CompositeHandler(
-      context.checkpointables_options.registry
-  )
-  handler_typestrs = {
-      name: handler_types.typestr(type(handler))
-      for name, handler in handler.get_handlers_for_save(
-          checkpointables
-      ).items()
-  }
-
-  return SaveResponse(
-      context.operation_id(),
-      tmp_path,
-      handler_typestrs,
-      background_awaitable,
-      start_time=start_time,
-      custom_metadata=custom_metadata,
-      context=context,
-      async_origin=async_origin,
-  )
-
-
-def check_directory_consistency(directory: path_types.PathLike):
-  """Raises an error if directory paths are not consistent across processes.
+def _check_directory_consistency(directory: path_types.PathLike):
+  """Raises error if directory paths are not consistent across processes.
 
   Args:
     directory: The directory path to check.
@@ -327,6 +330,39 @@ def check_directory_consistency(directory: path_types.PathLike):
     )
 
 
+class _TemporaryPathAwaitingCreation:
+  """A simple container for `PathAwaitingCreation` and `TemporaryPath`."""
+
+  def __init__(
+      self,
+      path: path_types.Path,
+      subdirectories: Iterable[str],
+      *,
+      use_snapshot: bool,
+  ):
+    self._temporary_path = saving_path_utils.get_temporary_path(
+        path, context=context_lib.get_context(), use_snapshot=use_snapshot
+    )
+    self._temporary_path_awaiting_creation = (
+        path_async_utils.PathAwaitingCreation.build(
+            self._temporary_path,
+            subdirectories,
+        )
+    )
+    assert (
+        self._temporary_path_awaiting_creation.path
+        == self._temporary_path.get()
+    )
+
+  @property
+  def path_awaiting_creation(self) -> path_async_utils.PathAwaitingCreation:
+    return self._temporary_path_awaiting_creation
+
+  @property
+  def temporary_path(self) -> atomicity_types.TemporaryPath:
+    return self._temporary_path
+
+
 def save_checkpointables_impl(
     path: path_types.PathLike,
     checkpointables: dict[str, Any],
@@ -337,43 +373,41 @@ def save_checkpointables_impl(
     partial_save: bool = False,
 ) -> async_types.AsyncResponse[None]:
   """See caller docstrings."""
+  start_time = time.time()
+  event_tracking.record_save_start(path, async_origin=async_origin)
   asyncio_utils.maybe_apply_nest_asyncio()
+  # Ensure the operation ID is incremented as soon as possible. This must be
+  # done uniquely for each save operation.
+  asyncio.run(context_lib.synchronize_next_operation_id())
   context = context_lib.get_context()
 
   path = context.file_options.path_class(path)
-  check_directory_consistency(path)
+  _check_directory_consistency(path)
   path_exists = path.exists() if partial_save else False
-
   # Prevent internal mutation from affecting the caller.
   checkpointables = dict(checkpointables)
-
-  start_time = time.time()
-  event_tracking.record_save_start(path, async_origin=async_origin)
-
-  tmp_path = saving_path_utils.get_temporary_path(
-      path, context=context, use_snapshot=path_exists
-  )
-
   checkpointables = add_internal_checkpointables(
       checkpointables, context=context
   )
   subdirectories = [] if path_exists else checkpointables.keys()
-
+  temporary_path = _TemporaryPathAwaitingCreation(
+      path,
+      subdirectories=subdirectories,
+      use_snapshot=path_exists,
+  )
   background_awaitable = asyncio.run(
-      run_blocking_save(
-          tmp_path,
+      _run_blocking_save(
+          temporary_path,
           checkpointables,
-          subdirectories,
           overwrite=overwrite,
           context=context,
           partial_save=partial_save,
       )
   )
-
-  return create_save_response(
+  return _SaveResponse.create(
       background_awaitable,
       checkpointables,
-      tmp_path,
+      temporary_path,
       start_time,
       context=context,
       custom_metadata=custom_metadata,