+++ checkpoint/orbax/checkpoint/experimental/v1/_src/serialization/compatibility.py
@@ -56,10 +56,10 @@ class _PathAwaitingCreation(path_types.PathAwaitingCreation):
     self._operation_id = operation_id
 
   def __truediv__(
-      self, other: path_types.PathAwaitingCreation | path_types.PathLike
+      self, other: path_types.PathLike
   ) -> path_types.PathAwaitingCreation:
-    if isinstance(other, path_types.PathAwaitingCreation):
-      other = other.path
+    if not isinstance(other, path_types.PathLike):
+      raise TypeError(f'Expected PathLike, got {type(other)}.')
     return _PathAwaitingCreation(self._path / other, self._operation_id)
 
   async def await_creation(self) -> path_types.Path:
@@ -410,11 +410,20 @@ class CompatibleTypeHandler(
     ):
       return self._leaf_handler._handler_impl._array_metadata_store  # pylint: disable=protected-access
     else:
-      logging.warning(
+      logging.info(
           'Cannot resolve _array_metadata_store for this v1 leaf handler: %r',
           self._leaf_handler,
       )
       return None
+
+  @property
+  def has_dispatcher(self) -> bool:
+    if hasattr(self._leaf_handler, '_handler_impl') and hasattr(
+        self._leaf_handler._handler_impl, 'has_dispatcher'  # pylint: disable=protected-access
+    ):
+      return self._leaf_handler._handler_impl.has_dispatcher  # pylint: disable=protected-access
+    else:
+      return False
 
 
 def get_v0_type_handler_registry(